Fix UB in bug() and let the compiler check print2 formats
bug() passed an int to a %ld conversion, which is undefined.  Print bugNum
with %d and drop the (long) casts and the %ld in BUG_CHECK_FORMAT.  The 2201L
in pushTempAlloc() shares the macro and follows; bufsiz in print2's bug #1528
diagnostic loses its cast too.

Declare print2 with PRINTF_FORMAT_ATTR so gcc and clang catch the next such
mismatch; it expands to nothing where the GNU attribute syntax is
unavailable.  mmfatl's fatalErrorPush is left alone: it implements its own
%s/%u-only format and allows unused trailing arguments, so the printf
archetype does not apply.

The attribute exposes 26 warnings, left in place here so they can be
inspected, and fixed in what follows.

+++ src/mmdata.c
@@ -753,11 +753,11 @@ void bug(int bugNum)
 
   if (mode == 2) {
     // If user chose to ignore bugs, print brief info and return
-    print2("?BUG CHECK:  *** DETECTED BUG %ld, IGNORING IT...\n", (long)bugNum);
+    print2("?BUG CHECK:  *** DETECTED BUG %d, IGNORING IT...\n", bugNum);
     return;
   }
 
-  print2(BUG_CHECK_FORMAT, (long)bugNum);
+  print2(BUG_CHECK_FORMAT, bugNum);
   if (mode == 0) { // Print detailed info for first bug
     print2("\n");
     print2("To get technical support, please open an issue \n");
@@ -765,7 +765,7 @@ void bug(int bugNum)
     print2("detailed command sequence or a command file that reproduces this bug,\n");
     print2("along with the source file that was used.  See HELP OPEN LOG for help on\n");
     print2("recording a session.  See HELP SUBMIT for help on command files.  Search\n");
-    print2("for \"bug(%ld)\" in the m*.c source code to find its origin.\n", bugNum);
+    print2("for \"bug(%d)\" in the m*.c source code to find its origin.\n", bugNum);
     print2("If earlier errors were reported, try fixing them first, because they\n");
     print2("may occasionally lead to false bug detection\n");
     print2("\n");
